SLING-8266 - Stop embedding ESAPI
Adapt tests by porting some of them to Sling Mocks. With the new
requirements added by the XSS API bundle, we need to pull in a more
recent version of the api bundle. With that, the Sling Testing Tools
mocks no longer provided all the needed methods so just switch to Sling
mocks outright.